Add clippy Makefile target, fix all clippy warnings, reject negative setpgid
- Add 'make clippy' target: lints kernel crates via custom bare-metal
  target, then host tools (xtask, z-config, region-alloc) natively
- Fix all xtask needless_borrows_for_generic_args (28 sites)
- Fix xtask unnecessary_map_or (is_some_and)
- Fix zCore/build.rs unnecessary_map_or (is_ok_and)
- Reject negative pgid in sys_setpgid (code review comment)

+++ linux-syscall/src/lib.rs
@@ -248,7 +248,7 @@ impl Syscall<'_> {
             Sys::SETUID => self.sys_setuid(a0 as u32),
             Sys::GETEUID => self.sys_geteuid(),
             Sys::GETEGID => self.sys_getegid(),
-            Sys::SETPGID => self.sys_setpgid(a0, a1 as u64),
+            Sys::SETPGID => self.sys_setpgid(a0, a1 as isize),
             Sys::GETPPID => self.sys_getppid(),
             Sys::SETSID => self.sys_setsid(),
             Sys::GETSID => self.sys_getsid(a0),
